auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
PyJWT
email-validator
python-multipart
bcrypt==4.0.1

